requirements.txt
boto3==1.9.155
botocore==1.12.155
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.3
Flask-Session==0.3.1
google-api-core==1.11.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-cloud-language==1.2.0
google-cloud-vision==0.36.0
googleapis-common-protos==1.6.0
grpcio==1.20.1
gunicorn==19.9.0
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
oauthlib==3.0.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.3
Werkzeug==0.15.4
